python-project/math-api/requirements.txt
fastapi
uvicorn
aiokafka
pydantic
pydantic-settings
python-dotenv
oracledb
pydantic-settings
httpx
-e ./shared_core